refactor src to generate only libflexflow.so
